+++ CotEditor/Sources/LayoutManager.swift
@@ -275,18 +275,6 @@ final class LayoutManager: NSLayoutManager, ValidationIgnorable {
     }
     
     
-    /// textStorage did update
-    override func processEditing(for textStorage: NSTextStorage, edited editMask: NSTextStorageEditActions, range newCharRange: NSRange, changeInLength delta: Int, invalidatedRange invalidatedCharRange: NSRange) {
-        
-        // invalidate wrapping line indent in editRange if needed
-        if editMask.contains(.editedCharacters) {
-            self.invalidateIndent(in: newCharRange)
-        }
-        
-        super.processEditing(for: textStorage, edited: editMask, range: newCharRange, changeInLength: delta, invalidatedRange: invalidatedCharRange)
-    }
-    
-    
     /// fill background rectangles with a color
     override func fillBackgroundRectArray(_ rectArray: UnsafePointer<NSRect>, count rectCount: Int, forCharacterRange charRange: NSRange, color: NSColor) {
         
@@ -324,76 +312,6 @@ final class LayoutManager: NSLayoutManager, ValidationIgnorable {
         let multiple = self.firstTextView?.defaultParagraphStyle?.lineHeightMultiple ?? 1.0
         
         return multiple * self.defaultLineHeight
-    }
-    
-    
-    /// invalidate indent of wrapped lines
-    func invalidateIndent(in range: NSRange) {
-        
-        assert(Thread.isMainThread)
-        
-        guard UserDefaults.standard[.enablesHangingIndent] else { return }
-        
-        guard
-            let textStorage = self.textStorage,
-            let textView = self.firstTextView
-            else { return assertionFailure() }
-        
-        // only on focused editor
-        if let window = textView.window, !self.layoutManagerOwnsFirstResponder(in: window) { return }
-        
-        let string = textStorage.string as NSString
-        let lineRange = string.lineRange(for: range)
-        
-        guard !lineRange.isEmpty else { return }
-        
-        let hangingIndent = self.spaceWidth * CGFloat(UserDefaults.standard[.hangingIndentWidth])
-        let regex = try! NSRegularExpression(pattern: "^[ \\t]+(?!$)")
-        
-        // get dummy attributes to make calculation of indent width the same as layoutManager's calculation (2016-04)
-        let defaultParagraphStyle = textView.defaultParagraphStyle ?? .default
-        let indentAttributes: [NSAttributedString.Key: Any] = {
-            let typingParagraphStyle = (textView.typingAttributes[.paragraphStyle] as? NSParagraphStyle)?.mutable
-            typingParagraphStyle?.headIndent = 1.0  // dummy indent value for size calculation (2016-04)
-            
-            return [.font: self.textFont,
-                    .paragraphStyle: typingParagraphStyle,
-                ].compactMapValues { $0 }
-        }()
-        
-        var cache = [String: CGFloat]()
-        
-        // process line by line
-        textStorage.beginEditing()
-        string.enumerateSubstrings(in: lineRange, options: .byLines) { (substring: String?, substringRange, enclosingRange, stop) in
-            guard let substring = substring else { return }
-            
-            var indent = hangingIndent
-            
-            // add base indent
-            let baseIndentRange = regex.rangeOfFirstMatch(in: substring, range: substring.nsRange)
-            if baseIndentRange.location != NSNotFound {
-                let indentString = (substring as NSString).substring(with: baseIndentRange)
-                if let width = cache[indentString] {
-                    indent += width
-                } else {
-                    let width = NSAttributedString(string: indentString, attributes: indentAttributes).size().width
-                    cache[indentString] = width
-                    indent += width
-                }
-            }
-            
-            // apply new indent only if needed
-            let paragraphStyle = textStorage.attribute(.paragraphStyle, at: substringRange.location, effectiveRange: nil) as? NSParagraphStyle
-            if indent != paragraphStyle?.headIndent {
-                let mutableParagraphStyle = (paragraphStyle ?? defaultParagraphStyle).mutable
-                mutableParagraphStyle.headIndent = indent
-                
-                textStorage.addAttribute(.paragraphStyle, value: mutableParagraphStyle, range: substringRange)
-            }
-        }
-        
-        textStorage.endEditing()
     }